refactor: Split out functionality from DynamicColor.
