+++ src/rust/ensogl/lib/core/src/display/shape/primitive.rs
@@ -3,7 +3,9 @@
 pub mod def;
 pub mod shader;
 pub mod system;
+pub mod style_watch;
 
 pub use def::*;
 pub use shader::*;
 pub use system::*;
+pub use style_watch::*;
