+++ src/rust/ide/view/graph-editor/src/component/node/port.rs
@@ -15,8 +15,8 @@ use ensogl::display::traits::*;
 use ensogl::display;
 use ensogl::gui::component;
 use ensogl::gui::cursor;
-use ensogl_shape_utils::dynamic_color::DynamicColor;
-use ensogl_shape_utils::dynamic_color;
+use ensogl_shape_utils::component_color::ComponentColor;
+use ensogl_shape_utils::component_color;
 use ensogl_text as text;
 use ensogl_text::buffer;
 use ensogl_theme as theme;
@@ -149,7 +149,7 @@ impl Manager {
         let type_color_map = default();
         let label          = app.new_view::<text::Area>();
         let ports          = default();
-        let text_color     = DynamicColor::new(&app);
+        let text_color     = ComponentColor::new(&app);
 
         // FIXME[WD]: Depth sorting of labels to in front of the mouse pointer. Temporary solution.
         // It needs to be more flexible once we have proper depth management.
@@ -191,9 +191,9 @@ impl Manager {
 
             eval set_dimmed ([text_color](should_dim) {
                 if *should_dim {
-                   text_color.frp.set_state(dynamic_color::State::Dim);
+                   text_color.frp.set_state(component_color::State::Dim);
                  } else {
-                   text_color.frp.set_state(dynamic_color::State::Base);
+                   text_color.frp.set_state(component_color::State::Base);
                  }
             });
         }
@@ -214,7 +214,7 @@ impl Manager {
 
         let text_color_path    = theme::vars::graph_editor::node::text::color;
         let text_color_path    = display::style::Path::from(text_color_path);
-        let text_dynamic_color = dynamic_color::Source::from(text_color_path);
+        let text_dynamic_color = component_color::Source::from(text_color_path);
         text_color.frp.set_source(text_dynamic_color);
 
         label.set_default_text_size(text::Size(12.0));
